requirements.txt
Flask
requests
gunicorn
fitbit
cherrypy
azure-storage-blob>=12.14.1
